+++ src/ThumbnailDot/index.types.ts
@@ -3,9 +3,9 @@ import type { GridProps } from '@mui/material'
 
 export type SpinoramaThumbnailDotProps = GridProps & {
 	/**
-	 * Indicates whether the dot is currently active.
-	 * Use this prop to style the component when it represents the active thumbnail.
+	 * Indicates whether the dot is currently selected.
+	 * Use this prop to style the component when it represents the selected thumbnail.
 	 * @default false
 	 */
-	active?: boolean
+	selected?: boolean
 }