+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Release Notes
 All notable changes to this package will be documented in this file.
 
+- Changed `active` to `selected` in the **ThumbnailDot** component
+
 # 0.4.3
 - Added `margin-top` to **Actions** component